**Major**
- Changed license from GPLv3 to AGPL with due permission from the All-father.

**Important**
- Fixed links not showing in /Kaizoku 
- Fixes and changes to secure bot from infinitely running regex.
- Fixes and improvements for blacklists and filters. 

**Minor**
- Removed a LOTR "You shall not pass" reference cause it wasnt sitting well with a sudents group.
Imagine joining a study group and seeing "you shall not pass" - kek. 
- Some translation updates and smaller fixes here and there. 
- Added more bugs to fix later on.

+++ tg_bot/modules/blacklist.py
@@ -12,14 +12,9 @@ from tg_bot.modules.disable import DisableAbleCommandHandler
 from tg_bot.modules.helper_funcs.chat_status import user_admin, user_not_admin, connection_status
 from tg_bot.modules.helper_funcs.extraction import extract_text
 from tg_bot.modules.helper_funcs.misc import split_message
+import tg_bot.modules.helper_funcs.regex_helper as r_helper
 
 BLACKLIST_GROUP = 11
-
-def infinite_loop_check(regex):
-     loop_matches = [r'\((.{1,}[\+\*]){1,}\)[\+\*].', r'[\(\[].{1,}\{\d(,)?\}[\)\]]\{\d(,)?\}', r'\(.{1,}\)\{.{1,}(,)?\}\(.*\)(\+|\* |\{.*\})']
-     for match in loop_matches:
-          match_1 = re.search(match, regex)
-          if match_1: return True
 
 @run_async
 @connection_status
@@ -77,7 +72,7 @@ def add_blacklist(bot: Bot, update: Update):
             except Exception as exce:
                 msg.reply_text(f"Couldn't add regex, Error: {exce}")
                 return
-            check = infinite_loop_check(trigger)
+            check = r_helper.infinite_loop_check(trigger)
             if not check:
                sql.add_to_blacklist(chat.id, trigger.lower())
             else:
@@ -147,14 +142,18 @@ def del_blacklist(bot: Bot, update: Update):
         return
 
     chat_filters = sql.get_chat_blacklist(chat.id)
+    error, match = False, False
     for trigger in chat_filters:
         pattern = r"( |^|[^\w])" + trigger + r"( |$|[^\w])"
-        try:
-          match = re.search(pattern, to_match, flags=re.IGNORECASE)
-        except Exception:
-          sql.rm_from_blacklist(chat.id, trigger)
-          msg.reply_text(f'Removed {trigger} from blacklist because of broken regex')
-          return
+        match = r_helper.regex_searcher(pattern, to_match)
+        if match == 'Timeout':
+           reason, error = match, True
+        elif match == 'Broken':
+            reason, error = 'Broken regex', True
+        if error:
+           sql.rm_from_blacklist(chat.id, trigger)
+           msg.reply_text(f'Removed {trigger} from blacklist because of {reason}')
+           return
         if match:
             try:
                 message.delete()
